% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/baad_data.R
\name{baad_data}
\alias{baad_data}
\alias{baad_delete}
\title{Load the baad database}
\usage{
baad_data(version = "ecology")

baad_delete(version = NULL)
}
\arguments{
\item{version}{Version to load.  "ecology" the published verison.
Other valid versions are "0.2.0", "0.9.0" and "1.0.0" which are
stored on github.  The "1.0.0" release corresponds to the
"ecology" release.}
}
\description{
Load the baad database.
}
\details{
The first time this is run for a given version, this function will
download the baad database, either from the Ecology website (with
version "ecology" or from github (numbered versions), unpack the
resulting zip file and load the csv files.  This might take a few
seconds to a minute.  Subsequent calls will be considerably
quicker because we cache both the downloaded data and the result
of reading the csv.  Subsequent calls after \emph{that} will be
essentially instantaneous.

The function \code{baad_delete} deletes all traces of downloaded
baad data if a version is not given, or a specific version if that
is listed.
}
\examples{
baad <- baad_data()
head(baad$data)
}

