% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baad_data.R
\name{baad_data_info}
\alias{baad_data_info}
\title{Github release information}
\usage{
baad_data_info(path = NULL)
}
\arguments{
\item{path}{Optional path in which to store the data.  If omitted
we use \code{rappdirs} to generate a reasonable path.}
}
\description{
Information to describe how to process github releases
}

