% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baad_data.R
\name{baad_data_versions}
\alias{baad_data_version_current}
\alias{baad_data_versions}
\title{Get release versions}
\usage{
baad_data_versions(local = TRUE, path = NULL)

baad_data_version_current(local = TRUE, path = NULL)
}
\arguments{
\item{local}{Should we return local (TRUE) or github (FALSE)
version numbers?  Github version numbers are pulled once per
session only.  The exception is for
\code{github_release_version_current} which when given
\code{local=TRUE} will fall back on trying github if there are
no local versions.}

\item{path}{Optional path in which to store the data.  If omitted
we use \code{rappdirs} to generate a reasonable path.##'}
}
\description{
Get release versions
}

