% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/baad_data.R
\name{baad_versions}
\alias{baad_versions}
\title{BAAD version}
\usage{
baad_versions(type)
}
\arguments{
\item{type}{Where to look for versions: options are \code{global},
which checks github and includes the published \code{ecology}
version, \code{github} which is the github releases only and
\code{local} which is versions downloaded to this computer.}
}
\description{
List known versions of the baad database.  The last version
returned is most recent.
}

