% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baad_data.R
\name{baad_data}
\alias{baad_data}
\alias{baad_data_del}
\alias{baad_data_version_current}
\alias{baad_data_versions}
\title{Load the baad database}
\usage{
baad_data(version = NULL, path = NULL)

baad_data_versions(local = TRUE, path = NULL)

baad_data_version_current(local = TRUE, path = NULL)

baad_data_del(version, path)
}
\arguments{
\item{version}{Version to load.  Verion "1.0.0" corresponds to the
version published in Ecology in 2015.  Other valid versions are
"0.1.0", "0.2.0" and "0.9.0" which are stored on github but are
of historical interest only.}

\item{local}{Logical indicating if local or github versions should
be polled.  With any luck, \code{local=FALSE} is a superset of
\code{local=TRUE}.  For \code{mydata_version_current}, if
\code{TRUE}, but there are no local versions, then we do check
for the most recent github version.}
}
\description{
Load the baad database.
}
\details{
The first time this is run for a given version, this function will
download the baad database from github, using numbered versions),
unpack the resulting zip file and load the csv files.  This might
take a few seconds to a minute.  Subsequent calls will be
considerably quicker because we cache both the downloaded data and
the result of reading the csv.  Subsequent calls after \emph{that}
will be essentially instantaneous.

The function \code{baad_delete} deletes all traces of downloaded
baad data if a version is not given, or a specific version if that
is listed.
}
\examples{
\dontrun{
baad <- baad_data()
head(baad$data)
}
}

