% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baad_data.R
\name{baad_data}
\alias{baad_data}
\alias{baad_data_del}
\title{Load the BAAD database}
\usage{
baad_data(version = NULL, path = NULL)

baad_data_del(version, path)
}
\arguments{
\item{version}{Version to load.  Verion "1.0.0" corresponds to the
version published in Ecology in 2015.  Other valid versions are
"0.1.0", "0.2.0" and "0.9.0" which are stored on github but are
of historical interest only.}

\item{path}{Optional path in which to store the data.  If omitted
we use \code{rappdirs} to generate a reasonable path.}
}
\description{
Load data from the Biomass and Allometry Database.
}
\details{
The first time this is run for a given version, this function will
download the Biomass and Allometry Database (BAAD) from github, using numbered versions),
unpack the resulting zip file and load the csv files.  This might
take a few seconds to a minute.  Subsequent calls will be
considerably quicker because we cache both the downloaded data and
the result of reading the csv.  Subsequent calls after \emph{that}
will be essentially instantaneous.

The function \code{baad_delete} deletes all traces of downloaded
BAAD data if a version is not given, or a specific version if that
is listed.
}
\examples{
\dontrun{
baad <- baad_data()
head(baad$data)
}
}

